Move: items 페이지 폴더에 components폴더를 추가
items 페이지를 구성하는 컴포넌트들을 components 폴더로 이동하여 해당
페이지의 레이아웃과 컴포넌트를 구분함
